deleteById 코드리팩토링, updateById 완료

+++ lib/controller/product_controller.dart
@@ -1,6 +1,9 @@
 import 'package:data_app/domain/product/product.dart';
 import 'package:data_app/domain/product/product_http_repository.dart';
+import 'package:data_app/main.dart';
+import 'package:data_app/views/components/my_alert_dialog.dart';
 import 'package:data_app/views/product/list/product_list_view_store.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Provider에 ProductController를 등록하는 것.
@@ -9,9 +12,12 @@ final productController = Provider<ProductController>((ref){
   return ProductController(ref);
 });
 
-// View는 Controller 에 요청한다.
+/**
+ * 컨트롤러 : 비지니스 로직 담당
+ */
 class ProductController {
 
+  final context = navigatorKey.currentContext!;
   final Ref _ref;
   ProductController(this._ref);
 
@@ -26,12 +32,21 @@ class ProductController {
   }
 
   void deleteById(int id) {
-    int result = _ref.read(productHttpRepository).deleteById(id);
-    if(result == 1){
-      _ref.read(productListViewStore.notifier).removeProduct(id);
+    if(id != 4){
+      int result = _ref.read(productHttpRepository).deleteById(id);
+      if(result == 1){
+        _ref.read(productListViewStore.notifier).removeProduct(id);
+      }
     }else{
-
+      showCupertinoDialog(context: context, builder: (context) => MyAlertDialog(msg: "삭제실패"));
     }
   }
+
+  void updateById(int id, Product productReqDto) {
+    Product productRespDto = _ref.read(productHttpRepository).updateById(id, productReqDto);
+    _ref.read(productListViewStore.notifier).updateProduct(id, productRespDto);
+  }
+
+
 
 }
